Moved tests to new project Refactored namespaces Deleted RectangleExtensions.cs Deleted IRectangleLayouter.cs and SpiralRectangleLayouter.cs Deleted SpiralRectangleLayouter_Should.cs
