+++ Homework/Homework.sln
@@ -2,6 +2,8 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{D88D6041-28E5-4E28-9114-9273816C112A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationTests", "TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj", "{64F6E990-FD28-4C82-A693-AED891395CDB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -12,5 +14,9 @@ Global
 		{D88D6041-28E5-4E28-9114-9273816C112A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D88D6041-28E5-4E28-9114-9273816C112A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D88D6041-28E5-4E28-9114-9273816C112A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64F6E990-FD28-4C82-A693-AED891395CDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64F6E990-FD28-4C82-A693-AED891395CDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64F6E990-FD28-4C82-A693-AED891395CDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64F6E990-FD28-4C82-A693-AED891395CDB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
