--- Homework/TagsCloudVisualization/Layouters/RectangleLayouters/IRectangleLayouter.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization.Layouters.RectangleLayouters
-{
-    internal interface IRectangleLayouter
-    {
-        public Rectangle PutNextRectangle(Size rectangleSize);
-    }
-}
