--- Homework/TagsCloudVisualization/Tests/SpiralRectangleLayouter_Should.cs
@@ -1,10 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace TagsCloudVisualization.Tests
-{
-    [TestFixture]
-    public class SpiralRectangleLayouter_Should
-    {
-        
-    }
-}
